+++ packages/react/test/specs/behaviors/alertBehavior-test.tsx
@@ -10,14 +10,4 @@ describe('AlertBehavior.ts', () => {
     const expectedResult = alertBehavior({ danger: true })
     expect(expectedResult.attributes.content.role).toEqual('alert')
   })
-
-  test('use defaultBehavior if success prop is defined', () => {
-    const expectedResult = alertBehavior({ success: true })
-    expect(expectedResult.attributes.content).toBeUndefined()
-  })
-
-  test('use defaultBehavior if info prop is defined', () => {
-    const expectedResult = alertBehavior({ info: true })
-    expect(expectedResult.attributes.content).toBeUndefined()
-  })
 })